Clippy part2

+++ src/date.rs
@@ -12,6 +12,7 @@ pub struct Date {
 }
 
 impl Date {
+    #[must_use]
     pub fn new(year: i16, month: u8, day: f64) -> Self {
         Self {
             year,
@@ -26,6 +27,7 @@ impl Date {
     }
 
     /// Create a `Date` object from values: year, month, day, seconds, minutes, seconds.
+    #[must_use]
     pub fn from_full_date(
         year: i16,
         month: u8,
@@ -46,12 +48,14 @@ impl Date {
         }
     }
 
+    #[must_use]
     pub fn to_julian_day(&self) -> JulianDay {
         let jd = get_julian_day(self);
         JulianDay::new(jd)
     }
 
     /// Returns the days interval between two dates
+    #[must_use]
     pub fn interval(&self, other: &Self) -> f64 {
         let self_jd = self.to_julian_day();
         let other_jd = other.to_julian_day();
@@ -64,15 +68,18 @@ impl Date {
     }
 
     /// Returns the day of the week of a calendar date (1 is Monday, 7 is Sunday)
+    #[must_use]
     pub fn week_day(&self) -> u8 {
         let jd = self.to_julian_day();
         ((jd.get_value() + 1.5_f64) as i32 % 7) as u8
     }
 
+    #[must_use]
     pub fn is_leap(&self) -> bool {
         false
     }
 
+    #[must_use]
     pub fn year_day(&self) -> u16 {
         // todo implement is_leap()
         let k = {
@@ -83,8 +90,8 @@ impl Date {
             }
         };
 
-        ((275 as u16 * self.month as u16) as f64 / 9.0_f64) as u16
-            - k * ((self.month + 9) as f64 / 12.0_f64) as u16
+        (f64::from(275_u16 * u16::from(self.month)) / 9.0_f64) as u16
+            - k * (f64::from(self.month + 9) / 12.0_f64) as u16
             + (self.day as u16)
             - 30
     }
